Minimap Woes
Minimap still isn't working. I tried replacing it with a red square instead, but that did nothing. It only replaces the current area (which is the error) with a red square. Don't know wheather I should bother to carrt on working on it or wheather I should go and work on something else first. That's probably a better idea.

If the next commit talks about working on the minimap, you know I failed